+++ src/lib/missing.c
@@ -49,7 +49,7 @@ void *memrchr(const void *s, int c, size_t n)
     const char *prev = NULL;
 
     for (const char *cur = s; cur <= end; prev = cur++) {
-        cur = (const char *)memchr(cur, c, (size_t)(ptrdiff_t)(end - cur));
+        cur = (const char *)memchr(cur, c, (size_t)(end - cur));
         if (!cur)
             break;
     }
